Don't start configuration without PuppetDB certificate
In J.1.0.0, Spinal Stack is masterless, but PuppetDB still remains on
the install-server.

However, to make it work properly, PuppetDB needs to be run under a
proxy with Apache2, with SSL enabled.
The certificate should be placed in /etc/ssl/certs/puppetdb.pem.
Without this certificate, PuppetDB can't run and the whole deployment
will fail.

This patch aims to:
* if the file exist on the install-server, ensure it has the right
  permissions (security)
* if the file does not exist, fail the configure.sh process and explain
  why with a pointer to the official documentation.

The certificate can be placed:
* manually at bootstrap
* in the environment directory (in etc/ssl/certs/puppetdb.pem)
* in tne YAML env file when running inception, with puppetdb_pem
* parameter.

All of this is documented on
http://spinalstack.enovance.com/en/latest/deploy/components/puppetdb.html

+++ configure.sh
@@ -182,9 +182,9 @@ EOF
     fi
 done
 
-######################################################################
-# Step 0: provision the puppet master and the certificates on the nodes
-######################################################################
+#######################################################################
+# Step 0: provision the nodes for Puppet configuration & certificates #
+#######################################################################
 
 if [ $STEP -eq 0 ]; then
     configure_hostname
@@ -214,6 +214,11 @@ if [ $STEP -eq 0 ]; then
           ssh $SSHOPTS $USER@$h.$DOMAIN sudo cp -r /tmp/data /etc/puppet/data
         fi
     done
+    if ! test -f /etc/ssl/certs/puppetdb.pem ; then
+      echo "/etc/ssl/certs/puppetdb.pem file is missing so PuppetDB cannot be configured."
+      echo "More documentation about it: http://spinalstack.enovance.com/en/latest/deploy/components/puppetdb.html"
+      exit 1
+    fi
     puppet apply /etc/puppet/modules/cloud/scripts/bootstrap.pp | tee  $LOGDIR/puppet-master.step0.log
     puppet apply -e 'include ::cloud::install::puppetdb::server' | tee  $LOGDIR/puppet-master.step0.log
     puppet apply -e 'include ::cloud::install::puppetdb::config' | tee  $LOGDIR/puppet-master.step0.log
